Use Yarn cache and don't cache node_modules when using Yarn
When using Yarn we only restore its cache but not node_modules; Yarn uses its
local cache and doesn't connect to the internet in such a case anyway and it's
quite fast but at the same time it's more stable to have fresh node_modules
copy in every build. We're not copying the cache as it can be large but instead
use it directly.